+++ example/android/app/src/main/kotlin/com/example/paginated_bloc_widget_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.paginated_bloc_widget_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()